test readme refactor
