+++ test/README.md
@@ -1,0 +1,18 @@
+### Organization of tests
+
+-   **Http request adapters**: ./adapter/adapter.spec
+-   **Api query building**: ./api/api.spec
+-   **Bundling of code**: ./bundle/bundle.spec
+-   **Chaining of statements in the same query**: ./chainable.spec
+-   **Compilation utils, expressions & parameter validation**: ./compilation/compilation.spec
+-   **Evaluators**: ./evaluator/evaluator.spec
+-   **Function & parameters**: ./functions/functions.spec
+-   **Tag queries**: ./query.spec
+-   **Specific statements**: ./standalone/standalone.spec
+-   **Statement interactions**: ./statement.spec
+-   **Overpass api status interpretation**: ./status.spec
+-   **Sanitizations in specific statements**: ./sanitization/usage.spec
+-   **Sanitization implementation**: ./sanitization/sanitization.spec
+-   **Root target state functionality**: ./target.spec
+-   **Overpass api result interpretation**: ./validator/validator.spec
+-   **Completenes (deprecations, etc)**: ./nonsense.spec
