+++ test/compilation/sanitization/index.ts
@@ -1,20 +1,20 @@
 import { describe } from "@jest/globals";
-import { sanitizationBooleanTests } from "./boolean";
-import { sanitizationBoundingBoxTests } from "./boundingbox";
-import { sanitizationDateTests } from "./date";
-import { sanitizationEnumTests } from "./enums";
-import { sanitizationGeoPosTests } from "./geopos";
-import { sanitizationNumberTests } from "./numbers";
-import { sanitizationRegExpTests } from "./regexp";
-import { sanitizationStringTests } from "./string";
+import { parametersBooleanTests } from "./boolean";
+import { parametersBoundingBoxTests } from "./boundingbox";
+import { parametersDateTests } from "./date";
+import { parametersEnumTests } from "./enums";
+import { parametersGeoPosTests } from "./geopos";
+import { parametersNumberTests } from "./numbers";
+import { parametersRegExpTests } from "./regexp";
+import { parametersStringTests } from "./string";
 
-export function sanitizationCompileTests() {
-	describe("Numbers", sanitizationNumberTests);
-	describe("Strings", sanitizationStringTests);
-	describe("RegExp", sanitizationRegExpTests);
-	describe("Dates", sanitizationDateTests);
-	describe("GeoPos", sanitizationGeoPosTests);
-	describe("BoundingBox", sanitizationBoundingBoxTests);
-	describe("Enums", sanitizationEnumTests);
-	describe("Boolean", sanitizationBooleanTests);
+export function parametersCompileTests() {
+	describe("Numbers", parametersNumberTests);
+	describe("Strings", parametersStringTests);
+	describe("RegExp", parametersRegExpTests);
+	describe("Dates", parametersDateTests);
+	describe("GeoPos", parametersGeoPosTests);
+	describe("BoundingBox", parametersBoundingBoxTests);
+	describe("Enums", parametersEnumTests);
+	describe("Boolean", parametersBooleanTests);
 }
