+++ test/test.spec.ts
@@ -1,5 +1,5 @@
 import "./checkConnection";
-//
+/** For information regarding tests see /test/README.md */
 import "./adapter/adapter.spec";
 import "./api/api.spec";
 import "./bundle/bundle.spec";
@@ -9,9 +9,10 @@ import "./evaluator/evaluator.spec";
 import "./functions/functions.spec";
 import "./nonsense.spec";
 import "./query.spec";
+import "./sanitization/sanitization.spec";
+import "./sanitization/usage.spec";
 import "./standalone/standalone.spec";
 import "./statement.spec";
 import "./status.spec";
-import "./stringSanitization.spec";
 import "./target.spec";
 import "./validator/validator.spec";
