+++ test/validator/validator.spec.ts
@@ -3,6 +3,8 @@ import { apiValidatorErrorTests } from "./error";
 import { apiValidatorJsonTests } from "./json";
 import { apiValidatorXmlTests } from "./xml";
 
+/** For information regarding tests see /test/README.md */
+
 describe("Validator", () => {
 	describe("Error", apiValidatorErrorTests);
 	describe("JSON", apiValidatorJsonTests);
